Librarification of the code
